// src/enums/StatType.ts
export enum StatType {
    HP = "hpFlat",
    ATK = "atkFlat",
    DEF = "defFlat",
    HPP = "hpPercentage",
    ATKP = "atkPercentage",
    DEFP = "defPercentage",
    SPD = "speed",
    CR = "critRate",
    CD = "critDamage",
    HEALING = "healing",
    EHR = "effectHitRate",
    ERES = "effectRes",
    ER = "energyRegeneration",
    BE = "breakEffect",
    QUANTUM = "quantumDamage",
    PHYSICAL = "physicalDamage",
    FIRE = "fireDamage",
    ICE = "iceDamage",
    LIGHTNING = "lightningDamage",
    WIND = "windDamage",
    IMAGINARY = "imaginaryDamage",
    NO_OPTION = "No option"
}